Refactor into lists_column_factories
