+++ cpp/src/column/column_factories.cpp
@@ -20,14 +20,12 @@
 #include <cudf/detail/nvtx/ranges.hpp>
 #include <cudf/dictionary/dictionary_factories.hpp>
 #include <cudf/fixed_point/fixed_point.hpp>
-#include <cudf/lists/column_factories.hpp>
+#include <cudf/lists/lists_column_factories.hpp>
 #include <cudf/scalar/scalar.hpp>
 #include <cudf/scalar/scalar_factories.hpp>
 #include <cudf/strings/detail/fill.hpp>
 #include <cudf/utilities/error.hpp>
 #include <cudf/utilities/traits.hpp>
-#include "cudf/copying.hpp"
-#include "cudf/types.hpp"
 
 #include <thrust/iterator/constant_iterator.h>
